Update patch set 1
Patch Set 1:

This opens up a big can of worms.

The find7a and find7s have different battery sizes.  Supposedly we can deal with this by removing the battery entry and let it be detected.

But the rest of the power profile will also be different because the two devices have different SoCs.  We don't have any actual data for these values so they are currently just placeholders.

Patch-set: 1
Label: Code-Review=0