Update patch set 4
Patch Set 4:

msm8974-common covers bacon and both find7 variants.

bacon does have a multi-color led.

find7a and find7s are both blue-only, as far as we can determine, though it does expose all three color nodes in sysfs.

so the current code and overlays do work as intended, though they are admittedly a bit misleading.

there is a known issue with the find7s notification led code blinking when it shouldn't and vice versa.  I'll be trying to track that down soon, but I expect that is just a coding issue in the lights module.

Patch-set: 4