Update patch set 4
Patch Set 4:

Ok, but wouldn't it be better to fix the overlays too so they are not confusing anymore?

I have already uploaded the needed changes:
http://review.cyanogenmod.org/70692
http://review.cyanogenmod.org/70693
http://review.cyanogenmod.org/70694

These changes disable custom colors for the battery led on find7.

I'm working on the settings to disable custom colors for the notification led too, since it's pointless on single color leds:
http://review.cyanogenmod.org/70688
http://review.cyanogenmod.org/70702

Patch-set: 4