Update patch set 1
Patch Set 1:

I'd hate to split builds based on one xml file.

I've heard that if the battery capacity is omitted, it can be determined at runtime.  Is there any truth to that?

Also, I believe Oppo is releasing a single ROM for both models, correct?  So we probably can't get actual numbers for each model without runnning it on a monsoon.

If all the above is correct, maybe we can just omit the battery capacity, take the ColorOS values, and call it good.

Patch-set: 1