Update patch set 1
Patch Set 1:

Yes, I also removed the qcom location apk.  This is shared with bacon so I think we just need to wait for bacon to be updated and we'll inherit the solution.

Ironically, enabling wifi seems to make it more stable (my initial flash was without gapps, so wifi was off).

Patch-set: 1