Update patch set 1
Patch Set 1: Code-Review+1

Metrodroid author here, just chiming in with some context to help you with the decision to merge this patch.

"com.nxp.mifare" is part of the Android CDD[0], C-4-1 and C-4-2.  This isn't the only device (or Android distribution) to have this problem -- eg: Google's factory Pixel 2, 3 and 3A images _also_ have this bug.

I added a workaround to Metrodroid so that it will report that the OPPO Find7 supports MIFARE Classic (despite the platform saying otherwise).  For the purposes of Metrodroid, it is 100% cosmetic, but it is confusing to users, because it will mark certain transit cards as "unsupported" if they use MIFARE Classic.  I like to err on the side of false _negatives_, to underpromise and overdeliver.

MIFARE Classic Tool used to also do this check, but now does something a whole lot more complicated[1].  Metrodroid also used to do this[2], but I found it had many false positives and false negatives with some vendors and third-party AOSP distributions doing strange things like including _multiple_ (ie: unused) NFC drivers (eg: both NXP and Broadcom).

The problem arrives with some other Android software designed for working with MIFARE Classic cards (exclusively) marks the "com.nxp.mifare" feature as "required" in the application's manifest, under the expectation that device vendors will follow the CDD[0].  This unfortunately prevents these applications from being installed (via Play Store) on otherwise compatible hardware.

I'm very keen for there to be one less device and Android distribution with this error, so I've marking with "+1", on the basis that this would do what it says on the tin.  I don't have the device in question, nor have I tried to compile this, but it looks the same as what is on other devices.


[0]: https://source.android.com/compatibility/9/android-9-cdd.html#7_4_4_near-field_communications
[1]: https://github.com/ikarus23/MifareClassicTool/blob/da0a92182f5be295351d4e0038c204ed3536601a/Mifare%20Classic%20Tool/app/src/main/java/de/syss/MifareClassicTool/Common.java#L576-L655
[2]: https://github.com/micolous/metrodroid/commit/5c487da2034577e5e7300cd446fb5abfa374a7cf

Patch-set: 1
Reviewer: Gerrit User 20848 <20848@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1